sentinel: organize workflow components
