+++ packages/sentinel/src/features/workflows/components/nodes/SchemaShape.tsx
