+++ packages/sentinel/src/pages/RunsPage.tsx
@@ -4,8 +4,8 @@ import { useInfiniteQuery, useQuery } from "@tanstack/react-query"
 import { History } from "lucide-react"
 import { useEffect, useMemo, useRef } from "react"
 import { LoadingInline, PageFrame } from "../components/common"
-import { RunHistoryFilters } from "../features/workflows/components/RunHistoryFilters"
-import { RunHistoryTable } from "../features/workflows/components/RunHistoryTable"
+import { RunHistoryFilters } from "../features/workflows/components/runs/RunHistoryFilters"
+import { RunHistoryTable } from "../features/workflows/components/runs/RunHistoryTable"
 import { useRunHistorySearch } from "../features/workflows/hooks/useRunHistorySearch"
 import { RUN_HISTORY_PAGE_SIZE } from "../features/workflows/utils/workflows"
